+++ core/save-xml.c
@@ -439,10 +439,6 @@ static void save_picture(struct membuffer *b, struct picture *pic)
 		put_degrees(b, pic->latitude, " gps='", " ");
 		put_degrees(b, pic->longitude, "", "'");
 	}
-	char *hash = hashstring(pic->filename);
-	if (!empty_string(hash))
-		put_format(b, " hash='%s'", hash);
-	free(hash);
 
 	put_string(b, "/>\n");
 }