+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Profile: Context menu entry to manually split a dive
+- Dive pictures: locate moved pictures based on filename and path
 - BLE support: fix recognition of Heinrich Weikamp OSTC 2 and OSTC Plus
 # Always add new entries at the very top of this file above other
 # existing entries and this note.